Arriba and XForm fixes for shim work.   - Add "sbyte" support in Arriba.   - Make AllCountResult and CountResult types public and move to Arriba.dll.   - Expose QueryIntelliSense.FilterToValidTablesForQuery for shim suggest.   - XForm: Make server port configurable.   - XForm: ReturnMethodItem will invoke any path prefix of URL (allowing routing)
